backend/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
python-decouple==3.8
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
